+++ external/lib/iarduino_Pressure_BMPMock/iarduino_Pressure_BMP.h
@@ -9,14 +9,13 @@ private:
     uint8_t _address;
 
 public:
-    float temperature;
-    float pressure;
-    float altitude;
+    double temperature;
+    double pressure;
+    double altitude;
 
     iarduino_Pressure_BMP(uint8_t address){};
 
     MOCK_METHOD(bool, begin, ());
-    // MOCK_METHOD(uint8_t, getAddress, ());
     MOCK_METHOD(bool, read, ());
     MOCK_METHOD(bool, read, (uint8_t)); // duplicate read because an argument is optional
 };
